adjust conditional compile & fix wx mode errir
