+++ packages/core/src/external/vue.js
