+++ packages/core/src/external/vuePlugin.js
