+++ packages/core/src/platform/patch/react/getDefaultOptions.android.js
@@ -1,0 +1,1 @@
+export { getDefaultOptions } from './getDefaultOptions.ios'
